Rename modules: 'pagination' -> 'paginator' and 'api_requests' -> 'session'
