+++ docs/reference.md
@@ -34,11 +34,11 @@ pyinaturalist.models
 :glob: true
 :maxdepth: 1
 
-modules/pyinaturalist.api_requests
 modules/pyinaturalist.auth
 modules/pyinaturalist.converters
 modules/pyinaturalist.exceptions
 modules/pyinaturalist.formatters
-modules/pyinaturalist.pagination
+modules/pyinaturalist.paginator
 modules/pyinaturalist.request_params
+modules/pyinaturalist.session
 ```
