+++ pyinaturalist/v0/observations.py
@@ -1,7 +1,6 @@
 from logging import getLogger
 from typing import List, Union
 
-from pyinaturalist.api_requests import delete, get, post, put
 from pyinaturalist.constants import (
     API_V0_BASE_URL,
     OBSERVATION_FORMATS,
@@ -13,8 +12,9 @@ from pyinaturalist.converters import convert_all_coordinates, convert_all_timest
 from pyinaturalist.docs import document_request_params
 from pyinaturalist.docs import templates as docs
 from pyinaturalist.exceptions import ObservationNotFound
-from pyinaturalist.pagination import add_paginate_all
+from pyinaturalist.paginator import add_paginate_all
 from pyinaturalist.request_params import convert_observation_params, validate_multiple_choice_param
+from pyinaturalist.session import delete, get, post, put
 
 logger = getLogger(__name__)
