+++ pyinaturalist/v1/projects.py
@@ -2,7 +2,7 @@ from pyinaturalist.constants import PROJECT_ORDER_BY_PROPERTIES, JsonResponse, M
 from pyinaturalist.converters import convert_all_coordinates, convert_all_timestamps
 from pyinaturalist.docs import document_request_params
 from pyinaturalist.docs import templates as docs
-from pyinaturalist.pagination import add_paginate_all
+from pyinaturalist.paginator import add_paginate_all
 from pyinaturalist.request_params import validate_multiple_choice_param
 from pyinaturalist.v1 import delete_v1, get_v1, post_v1
